% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCC_computation.R
\name{CCC_computation}
\alias{CCC_computation}
\title{CCC_computation}
\usage{
CCC_computation(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as x.}
}
\value{
A numeric value giving the CCC between x and y.
}
\description{
Compute the Concordance Correlation Coefficient (CCC) between two numeric vectors.
}
