% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_convertion.R
\name{color_convertion}
\alias{color_convertion}
\title{Convert numeric values to a color gradient from white to red.}
\usage{
color_convertion(x, max_scale = NULL)
}
\arguments{
\item{x}{A numeric vector to be converted to colors.}

\item{max_scale}{An optional maximum scale value for normalization.}
}
\value{
A vector of colors corresponding to the input values.
}
\description{
Convert numeric values to a color gradient from white to red.
}
