% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_pcf.R
\name{Compute_pcf}
\alias{Compute_pcf}
\title{Compute the pair correlation function using Fast Fourier Transform}
\usage{
Compute_pcf(
  sce,
  N_min_points = 50,
  r_vector = seq(0, 500, length.out = 100),
  computation_method = "direct",
  verbose = FALSE
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{N_min_points}{Minimum number of points required to compute the PCF for a ROI × cluster combination.}

\item{r_vector}{Numeric vector of radii at which to compute the PCF.}

\item{computation_method}{Either \code{"direct"}  or \code{"derivative"} The direct method calculates the pcf 
directly from the cellular point pattern model (pcf.ppp) using a kernel estimator,
which is conceptually simpler but generally slower for very dense point patterns. The derivative method uses the Fast Fourier Transform (FFT) 
to efficiently calculate Ripley's K function (Kest.fft), and then obtains the pcf by differentiating the K function for optimal performance on large datasets.}

\item{verbose}{Logical; whether to print progress information.}
}
\value{
A list containing:
  \item{List_pcf}{A list of PCF values for each ROI × cluster combination}
  \item{List_r}{A list of matching radius values}
  \item{Annotation}{A data.frame linking PCF entries to ROI and cluster}
}
\description{
Compute the pair correlation function using Fast Fourier Transform
}
