% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_sce_object.R
\name{Create_sce_object}
\alias{Create_sce_object}
\title{Create a SingleCellExperiment object from a data.frame}
\usage{
Create_sce_object(df, cell_centroid_x = 2, cell_centroid_y = 3, Labels = 4)
}
\arguments{
\item{df}{A data.frame containing at least coordinates and Labels info.}

\item{cell_centroid_x}{Name of the column for x coordinates (default: "cell_centroid_x").}

\item{cell_centroid_y}{Name of the column for y coordinates (default: "cell_centroid_y").}

\item{Labels}{Name of the column for cluster labels (default: "Labels").}
}
\value{
A SingleCellExperiment object with coordinates and labels.
}
\description{
Create a SingleCellExperiment object from a data.frame
}
