% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_beta_prime.R
\name{Fit_beta_prime}
\alias{Fit_beta_prime}
\title{Fits the beta prime model to Pcf(r)}
\usage{
Fit_beta_prime(x, y, show_plot = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as x.}

\item{show_plot}{Logical; whether to show fitted curve.}
}
\value{
A vector containing fitting parameters.
}
\description{
Fits the beta prime model to Pcf(r)
}
