% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_Dagum.R
\name{Fit_Dagum}
\alias{Fit_Dagum}
\title{Fits the Dagum model to Pcf(r)}
\usage{
Fit_Dagum(x, y, show_plot = FALSE, title = k)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as x.}

\item{show_plot}{Logical; whether to show fitted curve.}

\item{title}{Title for the plot.}
}
\value{
A vector containing fitting parameters.
}
\description{
Fits the Dagum model to Pcf(r)
}
