% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_generalized_beta_prime.R
\name{Fit_generalized_beta_prime}
\alias{Fit_generalized_beta_prime}
\title{Fits the generalized beta prime model to Pcf(r)}
\usage{
Fit_generalized_beta_prime(x, y, show_plot = FALSE, title = k)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as x.}

\item{show_plot}{Logical; whether to show the fitted curve.}

\item{title}{Title for the plot.}
}
\value{
A vector containing fitting parameters.
}
\description{
Fits the generalized beta prime model to Pcf(r)
}
