% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_log_Laplace.R
\name{Fit_log_Laplace}
\alias{Fit_log_Laplace}
\title{Fits the log laplace to Pcf(r)}
\usage{
Fit_log_Laplace(x, y, show_plot = FALSE, title = k)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as x.}

\item{show_plot}{Logical; whether to show the fitted curve.}

\item{title}{Title for the plot.}
}
\value{
A vector containing fitting parameters.
}
\description{
Fits the log laplace to Pcf(r)
}
