% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_mixed_interaction.R
\name{Fit_mixed_interaction}
\alias{Fit_mixed_interaction}
\title{Fit Mixed Interaction Model to PCF Data}
\usage{
Fit_mixed_interaction(x, y, show_plot = FALSE)
}
\arguments{
\item{x}{A numeric vector representing the distance values (r) for the PCF.}

\item{y}{A numeric vector representing the PCF values (\code{g(r)}) corresponding to \code{x}.}

\item{show_plot}{A logical value. If \code{TRUE}, a plot will be generated. Default is \code{FALSE}.}
}
\value{
A named numeric vector containing the fitted parameters and statistics.
}
\description{
Fit Mixed Interaction Model to PCF Data
}
