% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_parametric_pcf_model.R
\name{Fit_parametric_pcf_model}
\alias{Fit_parametric_pcf_model}
\title{Fit Parametric Models to the Pair Correlation Function (PCF)}
\usage{
Fit_parametric_pcf_model(List_pcf, model = "Sigmoid")
}
\arguments{
\item{List_pcf}{A list containing the PCF data to be fitted. It must contain:
\describe{
  \item{\code{List_pcf}}{A list of numeric vectors, where each vector is the g(r) value of the PCF.}
  \item{\code{List_r}}{A list of numeric vectors, where each vector is the corresponding distance r for the PCF.}
}}

\item{model}{A character string specifying the parametric model to fit. Supported options are:
\itemize{
  \item "Power_law"
  \item "Generalized_gamma"
  \item "Exponential"
  \item "Beta_prime"
  \item "Log_Cauchy"
  \item "Generalized_beta_prime"
  \item "Dagum"
  \item "Gamma"
  \item "Sigmoid" (default)
  \item "Log_Laplace"
}}
}
\value{
A \code{data.frame} containing the fitted parameters for each PCF in the input list.
}
\description{
Fit Parametric Models to the Pair Correlation Function (PCF)
}
