% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_power_law.R
\name{Fit_power_law}
\alias{Fit_power_law}
\title{Fits the power law model to Pcf(r)}
\usage{
Fit_power_law(x, y, show_plot = FALSE, title = k)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as x.}

\item{show_plot}{Logical; whether to show the fitted curve.}

\item{title}{Title for the plot.}
}
\value{
A vector containing fitting parameters.
}
\description{
Fits the power law model to Pcf(r)
}
