% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_sigmoid.R
\name{Fit_sigmoid}
\alias{Fit_sigmoid}
\title{Fits the sigmoid model to Pcf(r)}
\usage{
Fit_sigmoid(x, y, show_plot = FALSE, initial_tau = 100, title = k)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as x.}

\item{show_plot}{Logical; whether to show the fitted curve.}

\item{title}{Title for the plot.}
}
\value{
A vector containing fitting parameters.
}
\description{
Fits the sigmoid model to Pcf(r)
}
