% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_comparison_boxplot.R
\name{Model_comparison_boxplot}
\alias{Model_comparison_boxplot}
\title{Makes a boxplot comparing the R2 of different Pcf fitted model}
\usage{
Model_comparison_boxplot(
  list_results = list(Model_Sigmoid = Results_sigmoid, Model_exponential =
    Results_exponential)
)
}
\arguments{
\item{list_results}{A named list where each element is a \code{data.frame} or
a similar structure containing the fitting results for a specific model.
Each element **must** contain a column named \code{R2} storing the
coefficient of determination for each fit performed.}
}
\value{
NULL This function does not return a value; it displays a plot.
}
\description{
Makes a boxplot comparing the R2 of different Pcf fitted model
}
