% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_spatial_sampling_fitting.R
\name{Perform_spatial_sampling_fitting}
\alias{Perform_spatial_sampling_fitting}
\title{Perform spatial subsampling and fit parametric PCF models}
\usage{
Perform_spatial_sampling_fitting(
  sce,
  N_min_cells = 1000,
  FoV_width = 500,
  N_sampling = 500
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{N_min_cells}{Minimum number of cells required to compute the PCF.}

\item{FoV_width}{Width (in pixels or microns) of the square Field of View used for sampling.}

\item{N_sampling}{Number of spatial samplings to perform before filtering.}
}
\value{
The function returns the distribution of fitted parameters
across all valid sampled regions :
\describe{
  \item{Tau}{Matrix of \eqn{\tau} values across samplings.}
  \item{P}{Matrix of \eqn{p} values across samplings.}
  \item{C_normalised}{Matrix of normalised amplitudes.}
  \item{R2}{Matrix of model \eqn{R^2} values.}
}
}
\description{
Perform random spatial sampling on a \code{SingleCellExperiment} object, compute the PCF on each sampled region, 
and fit a parametric PCF model (Sigmoid).
}
