% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_selected_cluster.R
\name{Plot_selected_cluster}
\alias{Plot_selected_cluster}
\title{Plot cells of a selected cluster}
\usage{
Plot_selected_cluster(
  Meta_data = Meta_data,
  selected_cluster,
  title_show = NULL
)
}
\arguments{
\item{selected_cluster}{Cluster name or ID to plot.}

\item{title_show}{Optional title for the plot.}
}
\value{
A plot of the point pattern for the selected cluster.
}
\description{
Plot cells of a selected cluster
}
